+++ connector/src/test/java/it/eng/connector/integration/datatransfer/DataTransferAPIViewDataIntegrationTest.java
@@ -13,12 +13,14 @@ import java.io.ByteArrayInputStream;
 import java.time.Instant;
 import java.util.Arrays;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Stream;
 
 import org.bson.Document;
 import org.bson.types.ObjectId;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.http.HttpHeaders;
@@ -80,10 +82,11 @@ public class DataTransferAPIViewDataIntegrationTest extends BaseIntegrationTest{
 		policyEnforcementRepository.deleteAll();
 	}
 
-	@Test
 	@DisplayName("View data - success")
+	@ParameterizedTest
+	@MethodSource("getValidConstraints")
     @WithUserDetails(TestUtil.API_USER)
-	public void viewData_success() throws Exception {
+	public void viewData_success(Constraint constraint) throws Exception {
 		String datasetId = createNewId();
 		String fileContent = "Hello, World!";
 		
@@ -97,27 +100,7 @@ public class DataTransferAPIViewDataIntegrationTest extends BaseIntegrationTest{
 				.metadata(doc);
 		ObjectId fileId = gridFSBucket.uploadFromStream(FILE_NAME, new ByteArrayInputStream(fileContent.getBytes()), options);
 		
-		Agreement agreement = Agreement.Builder.newInstance()
-				.id(createNewId())
-				.assignee(NegotiationMockObjectUtil.ASSIGNEE)
-				.assigner(NegotiationMockObjectUtil.ASSIGNER)
-				.target(NegotiationMockObjectUtil.TARGET)
-				.timestamp(Instant.now().toString())
-				.permission(Arrays.asList(Permission.Builder.newInstance()
-						.action(Action.USE)
-						.constraint(Arrays.asList(Constraint.Builder.newInstance()
-								.leftOperand(LeftOperand.COUNT)
-								.operator(Operator.LTEQ)
-								.rightOperand("5")
-								.build()))
-						.build()))
-				.build();
-		
-		agreementRepository.save(agreement);
-		
-		PolicyEnforcement policyEnforcement = new PolicyEnforcement(createNewId(), agreement.getId(), 0);
-		
-		policyEnforcementRepository.save(policyEnforcement);
+		Agreement agreement = insertAgreement(constraint, 0);
 		
 		String consumerPid = createNewId();
 		String providerPid = createNewId();
@@ -148,7 +131,6 @@ public class DataTransferAPIViewDataIntegrationTest extends BaseIntegrationTest{
 		assertEquals("attachment;filename=" + FILE_NAME, result.andReturn().getResponse().getHeader(HttpHeaders.CONTENT_DISPOSITION));
 		assertEquals(response, fileContent);
 		
-		
 		// check if the TransferProcess is inserted in the database
 		TransferProcess transferProcessFromDb = transferProcessRepository.findById(transferProcessStarted.getId()).get();
 
@@ -163,39 +145,25 @@ public class DataTransferAPIViewDataIntegrationTest extends BaseIntegrationTest{
 		// check if the PolicyEnforcement count is increased
 		// waiting for 1 second to give time to the publisher to increase the policy access count
 		TimeUnit.SECONDS.sleep(1);
-		PolicyEnforcement enforcementFromDb = policyEnforcementRepository.findById(policyEnforcement.getId()).get();
-		
-		assertEquals(policyEnforcement.getCount() + 1, enforcementFromDb.getCount());
+		PolicyEnforcement enforcementFromDb = policyEnforcementRepository.findByAgreementId(agreement.getId()).get();
+		// increase count from initial 0 to 1
+		assertEquals(1, enforcementFromDb.getCount());
 		
 		ObjectId objectId = new ObjectId(transferProcessFromDb.getDataId());
 		gridFSBucket.delete(objectId);
     }
+	
+	private static Stream<Constraint> getValidConstraints() {
+	    return Stream.of(NegotiationMockObjectUtil.CONSTRAINT, NegotiationMockObjectUtil.CONSTRAINT_COUNT_5, NegotiationMockObjectUtil.CONSTRAINT_PURPOSE, 
+	    		NegotiationMockObjectUtil.CONSTRAINT_SPATIAL);
+	}
 
-	@Test
-	@DisplayName("View data - fail policy expired")
+	@ParameterizedTest
+	@MethodSource("getInvalidConstraints")
+	@DisplayName("View data - fail policy invalid")
     @WithUserDetails(TestUtil.API_USER)
-	public void viewData_failPolicyExpired() throws Exception {
-		Agreement agreement = Agreement.Builder.newInstance()
-				.id(createNewId())
-				.assignee(NegotiationMockObjectUtil.ASSIGNEE)
-				.assigner(NegotiationMockObjectUtil.ASSIGNER)
-				.target(NegotiationMockObjectUtil.TARGET)
-				.timestamp(Instant.now().toString())
-				.permission(Arrays.asList(Permission.Builder.newInstance()
-						.action(Action.USE)
-						.constraint(Arrays.asList(Constraint.Builder.newInstance()
-								.leftOperand(LeftOperand.COUNT)
-								.operator(Operator.LTEQ)
-								.rightOperand("5")
-								.build()))
-						.build()))
-				.build();
-		
-		agreementRepository.save(agreement);
-		
-		PolicyEnforcement policyEnforcement = new PolicyEnforcement(createNewId(), agreement.getId(), 6);
-		
-		policyEnforcementRepository.save(policyEnforcement);
+	public void viewData_fail_policyInvalid(Constraint constraint) throws Exception {
+		Agreement agreement = insertAgreement(constraint, 6);
 		
 		String consumerPid = createNewId();
 		String providerPid = createNewId();
@@ -228,5 +196,42 @@ public class DataTransferAPIViewDataIntegrationTest extends BaseIntegrationTest{
 		assertFalse(apiResp.isSuccess());
 		assertNull(apiResp.getData());
     }
+	
+	private static Stream<Constraint> getInvalidConstraints() {
+		Constraint constraintPurpose = Constraint.Builder.newInstance()
+				.leftOperand(LeftOperand.PURPOSE)
+				.operator(Operator.EQ)
+				.rightOperand("test")
+				.build();
+		
+		Constraint constraintSpatial = Constraint.Builder.newInstance()
+				.leftOperand(LeftOperand.SPATIAL)
+				.operator(Operator.EQ)
+				.rightOperand("USA")
+				.build();
+		
+	    return Stream.of(NegotiationMockObjectUtil.CONSTRAINT_DATEIME_INVALID, NegotiationMockObjectUtil.CONSTRAINT_COUNT_5, 
+	    		constraintPurpose, constraintSpatial);
+	}
 
+	private Agreement insertAgreement(Constraint constraint, int currentCount) {
+		Agreement agreement = Agreement.Builder.newInstance()
+				.id(createNewId())
+				.assignee(NegotiationMockObjectUtil.ASSIGNEE)
+				.assigner(NegotiationMockObjectUtil.ASSIGNER)
+				.target(NegotiationMockObjectUtil.TARGET)
+				.timestamp(Instant.now().toString())
+				.permission(Arrays.asList(Permission.Builder.newInstance()
+						.action(Action.USE)
+						.constraint(Arrays.asList(constraint))
+						.build()))
+				.build();
+		
+		agreementRepository.save(agreement);
+		
+		PolicyEnforcement policyEnforcement = new PolicyEnforcement(createNewId(), agreement.getId(), currentCount);
+		policyEnforcementRepository.save(policyEnforcement);
+		
+		return agreement;
+	}
 }